Began refactor
