+++ icortex/kernel/__init__.py
@@ -15,19 +15,13 @@ from icortex.context import ICortexContext
 from ipykernel.ipkernel import IPythonKernel
 from IPython import InteractiveShell, get_ipython
 from IPython.core.interactiveshell import ExecutionResult, ExecutionInfo
-from IPython.utils.io import capture_output
-from traitlets.config.configurable import SingletonConfigurable
 
 from icortex.helper import (
     escape_quotes,
-    yes_no_input,
-    highlight_python,
-    is_icortex_magic
+    is_icortex_magic,
 )
 from icortex.services import get_available_services
 from icortex.services.service_base import ServiceBase
-from icortex.services.service_interaction import ServiceInteraction
-from icortex.pypi import install_missing_packages, get_missing_modules
 from icortex.defaults import *
 from icortex.var import Var
 import importlib_metadata
@@ -74,7 +68,7 @@ class ICortexShell(InteractiveShell):
         # Initialize bound methods
         # TODO: make less hacky
         self.set_service = types.MethodType(ICortexShell.set_service, self)
-        self._run_dialog = types.MethodType(ICortexShell._run_dialog, self)
+        # self._run_dialog = types.MethodType(ICortexShell._run_dialog, self)
         self._check_service = types.MethodType(ICortexShell._check_service, self)
         self.print_service_help = types.MethodType(
             ICortexShell.print_service_help, self
@@ -82,7 +76,7 @@ class ICortexShell(InteractiveShell):
         self.run_cell = types.MethodType(ICortexShell.run_cell, self)
         self.cli = types.MethodType(ICortexShell.cli, self)
         self.prompt = types.MethodType(ICortexShell.prompt, self)
-        self.eval_prompt = types.MethodType(ICortexShell.eval_prompt, self)
+        # self.eval_prompt = types.MethodType(ICortexShell.eval_prompt, self)
         self.eval_var = types.MethodType(ICortexShell.eval_var, self)
         self.export = types.MethodType(ICortexShell.export, self)
         self.bake = types.MethodType(ICortexShell.bake, self)
@@ -90,76 +84,6 @@ class ICortexShell(InteractiveShell):
     def set_service(self, service: t.Type[ServiceBase]):
         self.service = service
         return True
-
-    def _run_dialog(
-        self,
-        code: str,
-        auto_execute: bool = DEFAULT_AUTO_EXECUTE,
-        auto_install_packages: bool = DEFAULT_AUTO_INSTALL_PACKAGES,
-        quiet: bool = DEFAULT_QUIET,
-        nonint: bool = False,
-    ) -> ServiceInteraction:
-
-        if not quiet:
-            print(highlight_python(code))
-
-        missing_modules = get_missing_modules(code)
-
-        install_packages_yesno = False
-        if len(missing_modules) > 0 and not auto_install_packages and not nonint:
-            install_packages_yesno = yes_no_input(
-                f"The following modules are missing in your environment: {', '.join(missing_modules)}\nAttempt to find and install corresponding PyPI packages?"
-            )
-        install_packages = auto_install_packages or install_packages_yesno
-
-        unresolved_modules = []
-        if install_packages:
-            # Unresolved modules are modules that cannot be mapped
-            # to any PyPI packages according to the local data in this library
-            unresolved_modules = install_missing_packages(code)
-            if len(unresolved_modules) > 0:
-                print(
-                    f"""The following imported modules could not be resolved to PyPI packages: {', '.join(unresolved_modules)}
-        Install them manually and try again.
-        """
-                )
-
-        # Modules that are still missing regardless of
-        # whether the user tried to auto-install them or not:
-        still_missing_modules = get_missing_modules(code)
-
-        execute_yesno = False
-        if not auto_execute and not nonint and len(still_missing_modules) == 0:
-            execute_yesno = yes_no_input("Proceed to execute?")
-
-        execute = auto_execute or execute_yesno
-
-        if execute and len(still_missing_modules) > 0:
-            execute = False
-            if auto_install_packages:
-                bermuda_modules = [
-                    module
-                    for module in still_missing_modules
-                    if module not in unresolved_modules
-                ]
-                print(
-                    f"""These modules should have been installed at this point, but they are still missing:  {', '.join(bermuda_modules)}
-    This might be due to an installer issue, please resolve manually.""",
-                )
-            print(
-                f"Skipping execution due to missing modules: {', '.join(still_missing_modules)}."
-            )
-
-        return ServiceInteraction(
-            name=self.service.name,
-            execute=execute,
-            outputs=[code],
-            quiet=quiet,
-            install_packages=install_packages,
-            missing_modules=missing_modules,
-            # unresolved_modules=unresolved_modules,
-            # still_missing_modules=still_missing_modules,
-        )
 
     def _check_service(self):
         if self.service is None:
@@ -217,7 +141,7 @@ class ICortexShell(InteractiveShell):
                 cell_id=cell_id,
             )
         elif input_type == InputType.PROMPT:
-            service_interaction = self.eval_prompt(raw_cell)
+            service_interaction = self.service.eval_prompt(raw_cell, self.history)
             code = service_interaction.get_code()
 
             result = InteractiveShell.run_cell(
@@ -311,10 +235,9 @@ class ICortexShell(InteractiveShell):
         return result
 
     def prompt(self, input_: str):
-        prompt = escape_quotes(input_)
         if self._check_service():
 
-            result = self.run_cell(prompt, input_type=InputType.PROMPT)
+            result = self.run_cell(input_, input_type=InputType.PROMPT)
         else:
             print(INIT_SERVICE_MSG)
 
@@ -322,28 +245,8 @@ class ICortexShell(InteractiveShell):
         prompt = escape_quotes(input_)
         eval_cli(prompt)
 
-    def eval_prompt(self, prompt_with_args: str) -> ServiceInteraction:
-        # Print help if the user has typed `/help`
-        argv = shlex.split(prompt_with_args)
-        args = self.service.prompt_parser.parse_args(argv)
-
-        # Otherwise, generate with the prompt
-        response = self.service.generate(
-            prompt_with_args,
-            context=self.history,
-            # context=self.history.get_dict(omit_last_cell=True),
-        )
-
-        # TODO: Account for multiple response values
-        code_: str = response[0]["text"]
-
-        return self._run_dialog(
-            code_,
-            auto_execute=args.execute,
-            auto_install_packages=args.auto_install_packages,
-            quiet=args.quiet,
-            nonint=args.nonint,
-        )
+    # def eval_prompt(self, prompt_with_args: str) -> ServiceInteraction:
+    #     return self.service.eval_prompt(prompt_with_args, self.history)
 
     def eval_var(self, line: str):
         "Evaluate var magic"
